Update player.h
fixed the value can never be less than what has been set to MAX_SPEED_ATTACKONFIST within config.lua

+++ src/creatures/players/player.h
@@ -2688,13 +2688,13 @@ class Player final : public Creature, public Cylinder {
 			if (onFistAttackSpeed) {
 				uint32_t baseAttackSpeed = vocation->getAttackSpeed();
 				uint32_t skillLevel = getSkillLevel(SKILL_FIST);
-				uint32_t attackSpeed = baseAttackSpeed - (skillLevel * g_configManager().getNumber(MULTIPLIER_ATTACKONFIST));
+				int32_t attackSpeed = baseAttackSpeed - (skillLevel * g_configManager().getNumber(MULTIPLIER_ATTACKONFIST));
 
 				if (attackSpeed < MAX_ATTACK_SPEED) {
 					attackSpeed = MAX_ATTACK_SPEED;
 				}
 
-				return attackSpeed;
+				return static_cast<uint32_t>(attackSpeed);
 			} else {
 				return vocation->getAttackSpeed();
 			}
